id:27 feat(signals-view): bug fixes
