+++ packages/ui/src/pages/ActivitiesPage/Activities/ImageGenerationDrawer/Image/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Image';
