+++ packages/ui/src/pages/ActivitiesPage/Activities/ImageGenerationDrawer/Signals/prettify-signals/index.ts
@@ -1,0 +1,1 @@
+export { default } from './prettify-signals';
