+++ packages/ui/src/pages/ActivitiesPage/Activities/ImageGenerationDrawer/Signals/prettify-signals/types.ts
@@ -1,0 +1,1 @@
+export type Output = [string, string][];
